+++ localstack/services/s3/constants.py
@@ -54,3 +54,15 @@ ALLOWED_HEADER_OVERRIDES = {
 # are creating bucket policies that enforce aws:SecureTransport, which makes the CDK deployment fail.
 # TODO: potentially look into making configurable
 ENABLE_MOTO_BUCKET_POLICY_ENFORCEMENT = False
+
+# params are required in presigned url
+SIGNATURE_V2_PARAMS = ["Signature", "Expires", "AWSAccessKeyId"]
+
+SIGNATURE_V4_PARAMS = [
+    "X-Amz-Algorithm",
+    "X-Amz-Credential",
+    "X-Amz-Date",
+    "X-Amz-Expires",
+    "X-Amz-SignedHeaders",
+    "X-Amz-Signature",
+]
